requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
multidict==6.0.5
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
